Combine subproject tests
